Update README.md
This project features European banking data and whether customers have subscribed to term deposits or certificates of deposit in the United States.  This data will be used to construct a model that will predict whether a customer will subscribe to a term deposit in the future.  The target is severely imbalanced.  Cost-sensitive methods will be explored that penalize the error of uncorrectly classifying the minority class. A cost-sensitive random forest, extra trees classifier, XGBoost, Balanced Random Forest from imblearn and the EasyEnsemble Classifier featuring AdaBoost will be used.  Ultimately, an artifical neural network will be utilized since they naturally calibrate probabilities.  A geometric mean scorer will be applied for this imbalanced class problem and a lift curve will be constructed which should exemplify whether the model is working as it should.